factor out code hash gen, db getter (note minor behavioral diff re: hint section with multiple snippets)
